% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.hash.R
\name{as.data.frame.hash}
\alias{as.data.frame.hash}
\title{Turn a hash into a data.frame}
\usage{
\method{as.data.frame}{hash}(x, ..., key = "key", value = "value")
}
\arguments{
\item{x}{hash}

\item{...}{additional arguments to \code{\link[=values]{values()}}}

\item{key}{string; name for the column of keys (Default: "key").}

\item{value}{string; name for the column of values (Default: "value")}
}
\value{
a data.frame with as many row as elements in \code{x} and two columns one for the
keys (always character) and value a list column
}
\description{
Transform a hash into a nrowx2 data.frame
}
\examples{

 h <- hash( a=1, b=1:2, c=1:3)
 as.data.frame(h)
 
}
