% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{attach}
\alias{attach}
\title{attach/detach a hash to the search path}
\usage{
attach(what, pos = 2L, ...)
}
\arguments{
\item{what}{hash; hash object to attach to the search path}

\item{pos}{integer where to attach the hash}

\item{...}{additional arguments passed to \code{\link[=base:attach]{base:attach()}}

\code{attach} attached the environment part of a hash to the search path.
The keys then become directly accessible.}
}
\description{
attach/detach a hash to the search path
}
\note{
this function is not-exported; explicit declaration is necessary
}
\examples{
  h <- hash( a=1, b=2, c=3 )
  hash:::attach( h )
  b                               # 2
  b <<- 25 
  b                               # 25
  h 
  detach(h)
  
}
\seealso{
\code{\link[=base:attach]{base:attach()}}
}
