% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_hash.R
\name{attach_hash}
\alias{attach_hash}
\alias{detach_hash}
\alias{attach}
\title{attach/detach a hash to the search path}
\usage{
attach_hash(hash, name = deparse(substitute(hash)), ...)

detach_hash(...)

attach(what, name = deparse(substitute(what)), ...)
}
\arguments{
\item{hash, what}{hash object to attach to the search path.}

\item{name}{string; name for the environment on the search path}

\item{...}{additional arguments passed to \code{\link[base:attach]{base::attach()}}

\code{attach_hash} attaches the  environment part of a hash to the search path.
The values become  directly accessible by their keys.

\code{hash:::attach()} is a non-exported versions of the same.}
}
\description{
attach/detach a hash to the search path
}
\examples{
  h <- hash( a=1, b=2, c=3 )
  
  attach_hash(h)
  search() 
  detach_hash(h)
  search()
  
  hash:::attach( h )
  b                               # 2
  b <<- 25 
  b                               # 25
  h 
  detach(h)
  
}
\seealso{
\itemize{
\item \code{\link[base:attach]{base::attach()}}
}
}
