% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\docType{methods}
\name{format}
\alias{format}
\alias{format,hash-method}
\title{format - format a hash object for print}
\usage{
\S4method{format}{hash}(x, max.print = getOption("max.print"), ...)
}
\arguments{
\item{x}{hash object}

\item{max.print}{maximum numbers of rows to print}

\item{...}{additional arguments}
}
\description{
Format a hash for printing/showing.
}
\seealso{
\code{\link[=format]{format()}}
}
