% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head.hash}
\alias{head.hash}
\alias{tail.hash}
\title{Head or tail of a hash}
\usage{
\method{head}{hash}(x, ...)

\method{tail}{hash}(x, ...)
}
\arguments{
\item{x}{hash;}

\item{...}{Additional arguments passed to \code{\link[utils:head]{utils::head()}}}
}
\value{
A hash object display the first \code{n} key-value pairs.
}
\description{
There is no guaranteed order to the keys-values are produced. This may change
in the futures.
}
\details{
Returns the first or last part of a hash
}
\examples{

  h <- hash( letters, 1:26)
  head(h)
  tail(h)
   
}
\seealso{
\itemize{
\item \code{\link[utils:head]{utils::head()}}
}
}
