% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\docType{methods}
\name{keys}
\alias{keys}
\alias{keys-methods}
\alias{keys,hash-method}
\alias{names,hash-method}
\alias{names}
\alias{names.hash}
\alias{keys<-}
\alias{keys<--methods}
\alias{keys<-,hash,ANY-method}
\alias{keys<-,hash,NULL-method}
\title{keys - get/set key(s) from a hash}
\usage{
keys(x, ...)

\S4method{keys}{hash}(x, sorted = getOption("hash.sorted", FALSE))

\S4method{names}{hash}(x)

keys(x) <- value

\S4method{keys}{hash,ANY}(x) <- value

\S4method{keys}{hash,ANY}(x) <- value

\S4method{keys}{hash,`NULL`}(x) <- value
}
\arguments{
\item{x}{A \code{\link[=hash]{hash()}} object.}

\item{...}{Used to allow for additional arguments for keys}

\item{sorted}{logical; whether the keys should be sorted
(DEFAULT: \code{getOption('hash.sorted', FALSE)})}

\item{value}{character or object coercable to character}
}
\value{
For \code{keys} and \code{names}, a character vector of key names
For the replacement methods \code{keys<-}, a hash object with the keys
renamed to \code{value}
}
\description{
Returns the key(s) from a hash, unsorted by default
}
\details{
Returns the character vector containing the keys of a hash object. By
default, the responses are not sorted. Set \code{sorted=TRUE} to return
keys in sort order.

\code{names} uses \code{base::names} and will always return a sorted list of
names. \code{keys} should generally be used in favor of \code{names}, but
names can be useful if you want sorted names for one instance when all others
don't provide a sorted list.
}
\examples{

  h <- hash( letters, 1:26 )
  keys(h)  # letters

  names(h) # same
  
  #' Rename keys
  # keys( h ) <- rev( keys(h) ) 

}
\seealso{
\code{\link[=setkeys]{setkeys()}} : rename specific hash keys.
\code{\link[=hash]{hash()}} : hash object
}
\author{
Christopher Brown
}
