% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_keys.R
\name{make_keys}
\alias{make_keys}
\alias{make.keys}
\title{creates/coerces objects to proper hash keys}
\usage{
make_keys(key)

make.keys(key)
}
\arguments{
\item{key}{An object that represents the key(s) to be coerced to a valid
hash keys.}
}
\value{
A character vector of valid keys/names
}
\description{
\strong{Notes:}
\itemize{
\item This function is not exported.
\item \code{make.keys()} will be deprecated in a future release; Use \code{make_keys()}
instead.
}
}
\details{
Given a vector of any type, \code{make_keys} tries to coerce it into a
character vector that can be used as a hash key. It is just a wrapper
around \code{\link[base:as.character]{base::as.character()}} with some additional error checking.

#' \code{make_keys} will replace \code{make.keys} in a future release.
}
\note{
This is used internally by the hash package and should not be normally
needed.  It is \emph{not} exported.
}
\examples{

\dontrun{
  make_keys( letters )
  make_keys( 1:26 )
}    
}
\seealso{
\itemize{
\item \code{\link[base:as.character]{base::as.character()}}
\item \code{\link[base:make.names]{base::make.names()}}
}
}
