% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.R
\docType{methods}
\name{del}
\alias{del}
\alias{delete}
\alias{del-methods}
\alias{delete-methods}
\alias{del,ANY,hash-method}
\alias{del,character,hash-method}
\alias{delete,ANY,hash-method}
\title{Remove key-value pair(s) from a hash}
\usage{
del(x, hash)

\S4method{del}{ANY,hash}(x, hash)

\S4method{del}{character,hash}(x, hash)

delete(x, hash)

\S4method{delete}{ANY,hash}(x, hash)
}
\arguments{
\item{x}{An object that will be coerced to valid key(s) to be removed from
the hash.  \code{x} will be coerced to a valid hash keys using
\code{\link[=make_keys]{make_keys()}}}

\item{hash}{A \code{\link[=hash]{hash()}} object}
}
\value{
None. This method exists solely for the side-effects of removing
items from the hash.
}
\description{
Removes key-value pair(s) from a hash by name of the object. There are also
R-like methods described in \link{Extract}. To delete all keys, use
\code{\link[=clear]{clear()}}.
}
\examples{

  h <- hash( letters, 1:26 )
  
  # USING del 
  del( "a", h )             # delete key  a
  del( c("b","c"), h )      # delete keys b, c
  
  # USING rm  
  rm( "d", envir=h )              # delete key  d
  rm( list= c("e","f"), envir=h ) # delete keys e,f
  
  # USING R syntsx
  h$g <- NULL               # delete key  g
  h[['h']] <- NULL          # delete key  h
  h['i'] <- NULL            # delete key  i
  h[ c('j','k')] <- NULL    # delete keys e,f
   
  keys(h)
  D
  
}
\seealso{
\itemize{
\item \code{\link[base:rm]{base::rm()}} base function used by \code{del}
\item \link{Extract} for R-like accessor
\item \code{\link[=clear]{clear()}} to remove all key-values and return an empty hash
\item \code{\link[=hash]{hash()}}
}
}
\author{
Christopher Brown
}
\keyword{data}
\keyword{manip}
\keyword{methods}
