\docType{methods}
\name{format}
\alias{format}
\alias{format,hash-method}
\title{format - format a hash object for prinint Methods for Function format in Package 'hash'}
\usage{
\S4method{format}{hash}(x, max.print = getOption("max.print"), ...)
}
\arguments{
  \item{x}{hash object}

  \item{max.print}{maximum numbers of rows to print}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
Format a hash for printing/showing.
}
\seealso{
\code{\link{format}}
}

