\name{Format hash object for pretty printing}

\docType{methods}

\alias{format,hash-method}
\alias{format}

\title{ Methods for Function format in Package 'hash' }

\description{
  Format a hash for printing. 
}

\section{Methods}{
\describe{
  \item{x = "hash"}{ Format a hash for pretty printing. }
}
}

\seealso{ See also \code{\link{format}} }

\keyword{methods}
